Organise classes in packages
