+++ app/src/main/AndroidManifest.xml
@@ -9,14 +9,14 @@
         android:icon="@mipmap/ic_launcher"
         android:theme="@style/AppTheme">
 
-        <activity android:name="fr.utc.assos.uvweb.ui.MainActivity">
+        <activity android:name="fr.utc.assos.uvweb.ui.activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
 
-        <activity android:name="fr.utc.assos.uvweb.ui.UvActivity" />
+        <activity android:name="fr.utc.assos.uvweb.ui.activity.UvActivity" />
     </application>
 
 </manifest>
