+++ app/src/main/java/fr/utc/assos/uvweb/ui/activity/ToolbarActivity.java
@@ -1,4 +1,4 @@
-package fr.utc.assos.uvweb.ui;
+package fr.utc.assos.uvweb.ui.activity;
 
 import android.os.Bundle;
 import android.support.v7.app.ActionBarActivity;
