+++ app/src/main/java/fr/utc/assos/uvweb/ui/adapter/NewsfeedAdapter.java
@@ -1,4 +1,4 @@
-package fr.utc.assos.uvweb.ui;
+package fr.utc.assos.uvweb.ui.adapter;
 
 import android.content.Context;
 import android.support.v7.widget.RecyclerView;
