+++ app/src/main/java/fr/utc/assos/uvweb/ui/view/SlidingTabLayout.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package fr.utc.assos.uvweb.ui;
+package fr.utc.assos.uvweb.ui.view;
 
 import android.content.Context;
 import android.graphics.Typeface;
